Move view model to its own file as it gets more complex.
